doctor_rest_crud

+++ one-health-rest/src/main/java/com/example/onehealthrest/service/impl/DepartmentServiceImpl.java
@@ -6,6 +6,7 @@ import com.example.onehealthcommon.mapper.DepartmentMapper;
 import com.example.onehealthcommon.repository.DepartmentRepository;
 import com.example.onehealthrest.service.DepartmentService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -14,45 +15,52 @@ import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class DepartmentServiceImpl implements DepartmentService {
     private final DepartmentRepository departmentRepository;
     private final DepartmentMapper departmentMapper;
 
-
     @Override
-    public Optional<Department> findById(int id) {
-        return departmentRepository.findById(id);
-    }
-
-    @Override
-    public Optional<Department> findByDepartments(String name) {
-        return departmentRepository.findByDepartments(name);
-    }
-
-    @Override
-    public Department save(Department department) {
-        return departmentRepository.save(department);
+    public DepartmentDto save(Department department) {
+        Optional<Department> byId = departmentRepository.findById(department.getId());
+        if (byId.isEmpty()) {
+            departmentRepository.save(department);
+            return departmentMapper.map(departmentRepository.save(department));
         }
-
+        log.info("add method save() did not work ");
+        return null;
+    }
 
     @Override
-    public List<DepartmentDto> fromDepartmentToDepartmentDto(List<Department> departments) {
-        List<DepartmentDto> departmentDto = new ArrayList<>();
-        for (Department dep : departments) {
-            departmentDto.add(departmentMapper.map(dep));
+    public DepartmentDto update(Department department) {
+        Optional<Department> byId = departmentRepository.findById(department.getId());
+        if (byId.isPresent()) {
+            Department departmentDb = byId.get();
+            departmentDb.setDepartments(departmentDb.getDepartments());
+            departmentRepository.save(departmentDb);
+            DepartmentDto departmentDto = departmentMapper.map(departmentDb);
+            return departmentDto;
         }
-        return departmentDto;
+        log.info(" method update() did not work ");
+        return null;
+    }
+
+    @Override
+    public List<Department> departmentList() {
+        return departmentRepository.findAll();
     }
 
 
     @Override
-    public boolean existsById(int id) {
-        return departmentRepository.existsById(id);
-    }
+    public boolean deleteById(int id) {
+        boolean isDeleted = false;
+        if (departmentRepository.existsById(id)) {
+            departmentRepository.deleteById(id);
+            log.info(" method deleteById() did not work ");
 
-    @Override
-    public void deleteById(int id) {
-        departmentRepository.deleteById(id);
+            return true;
+        }
+        return isDeleted;
     }
 
     @Override
